src/requirements.txt
black
codetiming
colorama
flake8-broken-line
flake8-bugbear
flake8-comprehensions
flake8-debugger
flake8-docstrings
flake8-isort
flake8-quotes
mypy
pytest
pytest-cov
pytest-flake8
python-decouple
requests